[ppo_balancer] Rename logging to callbacks
